requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
dataclass-factory==2.16
decorator==5.1.1
Deprecated==1.2.13
discord.py==2.0.1
ffmpeg-python==0.2.0
frozenlist==1.3.1
future==0.18.2
idna==3.4
iniconfig==1.1.1
multidict==6.0.2
mutagen==1.46.0
numpy==1.23.4
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycryptodomex==3.15.0
pydub==0.25.1
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
redis==4.3.4
shazamio==0.3.0.0
sniffio==1.3.0
tomli==2.0.1
typing_extensions==4.4.0
validators==0.20.0
websockets==10.3
wrapt==1.14.1
yarl==1.8.1
yt-dlp==2022.10.4
